Added 404, removed Node SASS - Making project too complicated
