+++ postcss.config.js
@@ -1,8 +1,6 @@
 module.exports = {
-	plugins: [
-		require('postcss-import'),
-		require('tailwindcss'),
-		require('tailwindcss/nesting')(require('postcss-nesting')),
-		require('autoprefixer'),
-	  ]
+	plugins: {
+		tailwindcss: {},
+		autoprefixer: {},
+	},
 };
