+++ public/assets/js/common.js
@@ -1,0 +1,1 @@
+
