--- public/js/common.js
@@ -1,1 +1,0 @@
-
