+++ src/helper/mail/templates/index.js
@@ -1,0 +1,7 @@
+const welcomeAndVerify = require("./welcomeAndVerify");
+const forgotPassword = require("./forgotPassword");
+
+module.exports = {
+    welcomeAndVerify,
+    forgotPassword,
+}
