--- src/helper/mailer/templates/index.js
@@ -1,7 +1,0 @@
-const welcomeAndVerify = require("./welcomeAndVerify");
-const forgotPassword = require("./forgotPassword");
-
-module.exports = {
-    welcomeAndVerify,
-    forgotPassword,
-}
