+++ views/partials/headerLinks.ejs
@@ -1,4 +1,12 @@
+<!-- Google Fonts - Nunito -->
+<link rel="preconnect" href="https://fonts.googleapis.com" />
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+<link
+	href="https://fonts.googleapis.com/css2?family=Nunito:ital,wght@0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,200;1,300;1,500;1,600;1,700;1,800;1,900&display=swap"
+	rel="stylesheet"
+/>
+
 <!-- Tailwind Output File -->
-<link rel="stylesheet" href="/assets/css/common.css">
+<link rel="stylesheet" href="/assets/css/common.css" />
 
 <!-- https://cdnjs.cloudflare.com/ajax/libs/validator/13.7.0/validator.min.js -->
