+++ views/serverError.ejs
@@ -1,17 +1,23 @@
 <%- include ("./partials/headerLoggedOut.ejs") %>
-<div class="serverError">
-	<h1 class="serverError__title"><%= error.message %></h1>
-	<h2 class="serverError__subtitle">
-		<% if (error.status === 404) { %> "<b><%= error.url %></b>" does not
-		exist in the server! <% } %>
-	</h2>
-	<div class="serverError__cta">
-		<img
-		src="/assets/images/404.gif"
-		class="serverError__image"
-		alt="There was an error!"
-		/>
-		<button class="serverError__cta-btn" onclick="window.location.replace('/')">Take me Home!</button>
+<div class="w-full h-full p-6">
+	<div class="flex items-center flex-col sm:flex-row justify-around w-4/5 sm:w-[60%] mx-auto transition-all duration-200">
+		<img src="/assets/images/404.gif" class="w-96 sm:w-full sm:max-w-[400px] h-auto" alt="There was an error!" />
+		<div class="flex flex-col gap-2 justify-center">
+			<h1 class="text-xl font-light"><%= error.message %></h1>
+			<h2 class="text-9xl text-gray-400 font-thin">404</h2>
+			<p class="text-3xl font-light">
+				Ooops! You weren't supposed to see this
+			</p>
+			<p class="text-lg">
+				Whatever you're looking for doesn't seem to exist. Which begs the question: <span class="text-gray-400">How certain are you that what you're looking for is actually what you need?!</span>
+			</p>
+			<a href="/" class="border border-gray-400 w-fit py-1 px-2 rounded hover:bg-gray-400 hover:text-white transition-all duration-200">
+				Take me Home!
+			</a>
+			<p class="font-light text-xs">
+				<% if (error.status === 404) { %> "<b><%= error.url %></b>" does not exist in the server! <% } %>
+			</h2>
+		</div>
 	</div>
 </div>
 <%- include ("./partials/footer.ejs") %>
